Setea fontsize por posicion
